--- session-service/src/main/java/com/sparta/moim/session/session/application/event/feign/OrganizationService.java
@@ -1,11 +1,0 @@
-package com.sparta.moim.session.session.application.event.feign;
-
-import com.sparta.moim.common.response.ApiResponseData;
-import com.sparta.moim.session.shared.enums.OrganizationMemberRole;
-import java.util.List;
-import java.util.UUID;
-
-public interface OrganizationService {
-  ApiResponseData<Boolean> checkRole(UUID organizationTrackingId, UUID userTrackingId,
-                                     List<OrganizationMemberRole> roles);
-}
